EVERYTHING IS WORKING!!!!!!
OKAYY SO
i made smth big 
no db=> cool! here is new one with data in it
if db is already created=> i unfrotunately didnt do this here where it would check for records and based on that add/ not add data for now, nothing happens
IM VERRY SORRY FOR not function-pretty short code BUT I TRIED OK? 😭  IT JUST DIDNT WORK THAT WAY so it is ugly longass BUT it works so. Maybe when i will be back from skiing ima optimize it BUT rn lets be happy that everything works nicely :)))